+++ sdk/source/data/port/port-in-out.cpp
@@ -2,7 +2,7 @@
 
 namespace adam 
 {
-    port_in_out::port_in_out(const string_hashed& item_name, size_t state_buffer_size) 
+    port_in_out::port_in_out(const string_hashed& item_name, uint32_t state_buffer_size) 
      :  port_input(item_name, state_buffer_size)
     {
